Extract ClientListTableRow into it's own component
